Renamed LocationMarker to UserMarker and moved it to its own component file.
