+++ client/Components/Map.jsx
@@ -1,7 +1,8 @@
 import { makeStyles } from "@material-ui/core";
-import React, {useEffect, useState} from "react";
+// import React, {useEffect, useState} from "react";
 // import { makeStyles } from "@material-ui/core/styles";
 import { MapContainer, TileLayer, Marker, Popup, useMapEvents } from 'react-leaflet';
+import UserMarker from './userMarker.jsx';
 import ToiletCard from './toiletCard.jsx';
 
 // toilet dependency injection goes here:
@@ -28,26 +29,26 @@ export default function Map() {
 
   const classes = useStyles();
 
-  const [coords, setCoords] = useState([40.785091, -73.968285]);
+  // const [coords, setCoords] = useState([40.785091, -73.968285]);
 
-  function LocationMarker() {
-    const map = useMapEvents({
-      click() {
-        map.locate()
-      },
-      locationfound(e) {
-        setCoords(e.latlng)
-        // map.flyTo(e.latlng, map.getZoom())
-        map.flyTo(e.latlng, map.getZoom())
-      },
-    })
+  // function UserMarker() {
+  //   const map = useMapEvents({
+  //     click() {
+  //       map.locate()
+  //     },
+  //     locationfound(e) {
+  //       setCoords(e.latlng)
+  //       // map.flyTo(e.latlng, map.getZoom())
+  //       map.flyTo(e.latlng, map.getZoom())
+  //     },
+  //   })
   
-    return coords === null ? null : (
-      <Marker position={coords}>
-        <Popup>You are here</Popup>
-      </Marker>
-    )
-  };
+  //   return coords === null ? null : (
+  //     <Marker position={coords}>
+  //       <Popup>You are here</Popup>
+  //     </Marker>
+  //   )
+  // };
   
   // useEffect(() => {
   //   navigator.geolocation.getCurrentPosition(function(position) {
@@ -63,11 +64,11 @@ return (
     attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
     url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
   />
-  <LocationMarker position={coords}>
+  <UserMarker position={coords}>
     <Popup>
       {/* A pretty CSS3 popup. <br /> Easily customizable. */}
       <ToiletCard />
     </Popup>
-  </LocationMarker>
+  </UserMarker>
 </MapContainer>
 )}
